Use require.resolve to serve livereload-js
The previous implementation depended on livereload-js being installed
into the 'node_modules' directory of this project.

On more complex projects with multiple dependencies that depend on
livereload-js, NPM will install livereload-js to the root node_modules
directory, resulting in the fixed path used here to be incorrect.

Node supplies require.resolve as a method to determine the location
of a dependency script.  This commit replaces the hard-coded path with
require.resolve to ensure we load the script from wherever it is
installed.